added useMutation to add users
