+++ Poke/src/routes.tsx
@@ -1,8 +1,19 @@
+import {ApolloProvider} from '@apollo/client';
+import React from 'react';
 import {Navigation} from 'react-native-navigation';
 import App from '../App';
-import Users from './pages/users';
+import {client} from './apollo-client';
 import {AddUser} from './pages/add-user';
+import Users from './pages/users';
 
 Navigation.registerComponent('Login', () => App);
 Navigation.registerComponent('Users', () => Users);
-Navigation.registerComponent('Add User', () => AddUser);
+Navigation.registerComponent('Add User', () => {
+  return (props) => {
+    return (
+      <ApolloProvider client={client}>
+        <AddUser {...props} />
+      </ApolloProvider>
+    );
+  };
+});
